Move walk_key_path() to uwtools.config.support
