+++ src/uwtools/config/jinja2.py
@@ -11,7 +11,7 @@ from typing import Optional, Union
 from jinja2 import Environment, FileSystemLoader, StrictUndefined, Undefined, meta
 from jinja2.exceptions import UndefinedError
 
-from uwtools.config.support import UWYAMLConvert, UWYAMLRemove, format_to_config
+from uwtools.config.support import UWYAMLConvert, UWYAMLRemove, format_to_config, walk_key_path
 from uwtools.logging import INDENT, MSGWIDTH, log
 from uwtools.utils.file import get_file_format, readable, writable
