+++ src/uwtools/config/support.py
@@ -68,6 +68,29 @@ def log_and_error(msg: str) -> Exception:
     return UWConfigError(msg)
 
 
+def walk_key_path(config: dict, key_path: list[str]) -> tuple[dict, str]:
+    """
+    Navigate to the sub-config at the end of the path of given keys.
+
+    :param config: A config.
+    :param key_path: Path of keys to subsection of config file.
+    :return: The sub-config and a string representation of the key path.
+    """
+    keys = []
+    pathstr = "<unknown>"
+    for key in key_path:
+        keys.append(key)
+        pathstr = ".".join(keys)
+        try:
+            subconfig = config[key]
+        except KeyError as e:
+            raise log_and_error(f"Bad config path: {pathstr}") from e
+        if not isinstance(subconfig, dict):
+            raise log_and_error(f"Value at {pathstr} must be a dictionary")
+        config = subconfig
+    return config, pathstr
+
+
 def yaml_to_str(cfg: dict) -> str:
     """
     Return a uwtools-conventional YAML representation of the given dict.
